updating date picker to fit input themes
