+++ src/components/basic/date-picker/ppDatePicker.scss
@@ -1,0 +1,7 @@
+.pp-date-picker {
+  padding: 15px 0px;
+
+  input {
+    max-width: 93%;
+  }
+}
