--- src/components/date-picker/datePicker.scss
@@ -1,26 +1,0 @@
-.pp-date-picker {
-  padding: 5px;
-  border: none;
-  border-radius: 5px;
-  outline: none;
-  font-size: 11px;
-  color: #676767;
-
-  &-container {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    border: 1px solid map-get($pp-colours, "grey");
-    padding: 3px;
-    margin: 0px 45px;
-    border-radius: 5px;
-  }
-
-  &-label {
-    color: #676767;
-    font-size: 10px;
-    margin-right: auto;
-    margin-left: 10px;
-  }
-}
